requirements.txt
# Flask and Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data Processing - Python 3.13 compatible
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1  # optional, for pandas future-proofing

# Machine Learning
scikit-learn==1.3.2
imbalanced-learn==0.12.0

# Visualization
matplotlib==3.5.3
seaborn==0.13.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Database - MongoDB
pymongo>=4.5.0
python-dotenv>=1.0.0

# Optional: Async MongoDB driver
# motor>=3.3.0
